# backend/openapi.yaml
openapi: 3.0.1
info:
  title: Home KM API
  description: "Home KM REST API. All authenticated endpoints require an `Authorization:\
    \ Bearer <jwt>` header."
  version: v1
servers:
- url: /
  description: Same-origin (proxied via the frontend nginx)
security:
- bearerAuth: []
paths:
...
  /api/folders/{id}/child-safe:
    put:
      tags:
      - folder-controller
      operationId: setChildSafe
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: boolean
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FolderResponse'
...
  /api/admin/users/{id}:
    put:
      tags:
      - admin-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      tags:
      - admin-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/admin/users:
    get:
      tags:
      - admin-controller
      operationId: listUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      tags:
      - admin-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/admin/users/{id}/sessions/revoke:
    post:
      tags:
      - admin-controller
      operationId: revokeUserSessions
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/admin/users/{id}/reset-password:
    post:
      tags:
      - admin-controller
      operationId: resetPassword
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
      responses:
        "200":
          description: OK
  /api/admin/users/bulk:
    post:
      tags:
      - admin-controller
      operationId: bulkImport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkImportRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BulkImportResult'
  /api/admin/review-queue/note/{id}:
    post:
      tags:
      - admin-review-controller
      operationId: reviewNote
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
        required: true
      responses:
        "200":
          description: OK
  /api/admin/review-queue/file/{id}:
    post:
      tags:
      - admin-review-controller
      operationId: reviewFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
        required: true
      responses:
        "200":
          description: OK
...
  /api/auth/sessions:
    get:
      tags:
      - auth-controller
      operationId: sessions
      parameters:
      - name: X-Refresh-Token
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
...
components:
  schemas:
...
    Crumb:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    FolderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        ownerId:
          type: integer
          format: int64
        isChildSafe:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
        color:
          type: string
        icon:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/FolderResponse'
        ancestors:
          type: array
          items:
            $ref: '#/components/schemas/Crumb'
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        displayName:
          type: string
        isAdmin:
          type: boolean
        isChild:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          maxLength: 100
          minLength: 1
          pattern: "^[^\\p{Cntrl}]+$"
          type: string
        isAdmin:
          type: boolean
        isChild:
          type: boolean
        isActive:
          type: boolean
...
    CreateUserRequest:
      required:
      - displayName
      - email
      - password
      type: object
      properties:
        email:
          maxLength: 320
          minLength: 0
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$"
          type: string
        displayName:
          maxLength: 100
          minLength: 1
          pattern: "^[^\\p{Cntrl}]+$"
          type: string
        isAdmin:
          type: boolean
        isChild:
          type: boolean
    ResetPasswordRequest:
      required:
      - newPassword
      type: object
      properties:
        newPassword:
          maxLength: 2147483647
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$"
          type: string
    BulkImportRequest:
      required:
      - rows
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/BulkRow'
    BulkRow:
      required:
      - displayName
      - email
      type: object
      properties:
        email:
          type: string
        displayName:
          maxLength: 100
          minLength: 0
          type: string
        isAdmin:
          type: boolean
        isChild:
          type: boolean
    BulkImportError:
      type: object
      properties:
        email:
          type: string
        reason:
          type: string
    BulkImportResult:
      type: object
      properties:
        created:
          type: integer
          format: int32
        skipped:
          type: integer
          format: int32
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BulkImportError'
    ReviewRequest:
      type: object
      properties:
        childSafe:
          type: boolean
...
    SessionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        deviceLabel:
          type: string
        userAgent:
          type: string
        ipAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        rememberMe:
          type: boolean
        current:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT